fail license and cache requests silently
