+++ checkov/common/bridgecrew/vulnerability_scanning/image_scanner.py
@@ -117,27 +117,33 @@ class ImageScanner:
     @staticmethod
     def get_scan_results_from_cache(image_id: str) -> Dict[str, Any]:
         image_id_encode = quote_plus(image_id)
-        response = request_wrapper(
-            "GET", f"{bc_integration.api_url}/api/v1/vulnerabilities/scan-results/{image_id_encode}",
-            headers=bc_integration.get_default_headers("GET"), should_call_raise_for_status=True
-        )
-        response_json = response.json()
-        output_type = response_json["outputType"]
 
-        logging.info(f"{output_type} returned from cache")
-
-        if output_type == "Error":
-            logging.error(response_json["outputData"])
-            return {}
-
-        elif output_type == "Result":
-            result: dict[str, Any] = json.loads(
-                decompress_file_gzip_base64(
-                    response_json["outputData"]
-                )
+        try:
+            response = request_wrapper(
+                "GET", f"{bc_integration.api_url}/api/v1/vulnerabilities/scan-results/{image_id_encode}",
+                headers=bc_integration.get_default_headers("GET"), should_call_raise_for_status=True
             )
-            return result
-        logging.info(f"Got an empty result for image={image_id}")
+            response_json = response.json()
+            output_type = response_json["outputType"]
+
+            logging.info(f"{output_type} returned from cache")
+
+            if output_type == "Error":
+                logging.error(response_json["outputData"])
+                return {}
+
+            elif output_type == "Result":
+                result: dict[str, Any] = json.loads(
+                    decompress_file_gzip_base64(
+                        response_json["outputData"]
+                    )
+                )
+                return result
+
+            logging.info(f"Got an empty result for image={image_id}")
+        except Exception:
+            logging.info("Failed to retrieve image cache result", exc_info=True)
+
         return {}
 
     def should_download(self) -> bool:
